+++ rubato/misc.py
@@ -18,7 +18,9 @@ def wrap(
     pos: Vector | tuple[float, float] = (0, 0),
     rotation: float = 0,
     z_index: int = 0,
-    debug: bool = False
+    debug: bool = False,
+    active: bool = True,
+    hidden: bool = False,
 ):
     """
     Wraps a component or list of components in a GameObject.
@@ -30,6 +32,8 @@ def wrap(
         rotation: The rotation of the GameObject. Defaults to 0.
         z_index: The z_index of the GameObject. Defaults to 0.
         debug: Whether the GameObject is in debug mode. Defaults to False.
+        active: Whether the GameObject is active. Defaults to True.
+        hidden: Whether the GameObject is hidden. Defaults to False.
 
     Raises:
         TypeError: If comp is not a Component or a list of Components.
@@ -37,7 +41,15 @@ def wrap(
     Returns:
         The wrapped GameObject.
     """
-    go = GameObject(name=name, pos=Vector.create(pos), rotation=rotation, z_index=z_index, debug=debug)
+    go = GameObject(
+        name=name,
+        pos=Vector.create(pos),
+        rotation=rotation,
+        z_index=z_index,
+        debug=debug,
+        active=active,
+        hidden=hidden
+    )
     if isinstance(comp, Component):
         go.add(comp)
     elif isinstance(comp, list):