+++ rubato/struct/gameobject/ui/slider.py
@@ -24,6 +24,7 @@ class Slider(Component):
         offset: The offset of the component from the game object. Defaults to Vector(0, 0).
         rot_offset: The rotation offset of the component from the game object. Defaults to 0.
         z_index: The z-index of the component. Defaults to 0.
+        hidden: Whether the slider is hidden. Defaults to False.
     """
 
     def __init__(
@@ -38,9 +39,10 @@ class Slider(Component):
         onexit: Callable | None = None,
         offset: Vector = Vector(),
         rot_offset: float = 0,
-        z_index: int = 0
+        z_index: int = 0,
+        hidden: bool = False
     ):
-        super().__init__(offset=offset, rot_offset=rot_offset, z_index=z_index)
+        super().__init__(offset=offset, rot_offset=rot_offset, z_index=z_index, hidden=hidden)
         self.dims: Vector = Vector(button_width, button_height)
         """The dimensions of the button."""
         self.pressed: bool = False